Move in_organisation trait into user factory
Although top level traits (mixins) are allowed according to the docs[1],
they appear to be broken in factory bot 6.4.0[2].

We were seeing the following error:

NoMethodError: undefined method `<' for nil:NilClass
  if for_class < ActiveRecord::Base && for_class.primary_key == attribute.name.to_s

When running the tests that used the `in_organisation` trait.

Given that we're only using this trait in User factories I've moved it
inside the User factory definition so that we can upgrade.

[1]: https://thoughtbot.github.io/factory_bot/traits/mixins.html
[2]: https://github.com/thoughtbot/factory_bot_rails/issues/431