Extract Account::UserApplicationPermissionPolicy
The `PermissionsController#show` action lists all permissions for the
application along with whether the user has that permission.

The `UserApplicationPermission` model provides us with the join between
Users and Applications via the SupportedPermission model so it feels
more natural to have a policy for this model which determines whether a
user can view these permissions.

At present this policy only takes the current user's role into account
but the plan is to generalise it so that we can apply it irrespective of
whether you're trying to view your own permissions, or permissions for
another user.