Move edit user organisation page to GOV.UK Design System
Trello: https://trello.com/c/fA6dRVSm

c.f. this commit [1] where we did the same for user name.

This changes the edit user organisation page to use the Design System
and makes a few other tweaks to bring it into line with the latest
designs.

The page title & heading are slightly different to match other similar
pages.

I've added breadcrumbs to make this page consistent with other pages
even though they're not shown in the latest design for this type of
page.

Only Admins & Superadmins can change another user's organisation and so
the "Change (organisation)" link is only displayed alongside the user's
organisation name on the "edit user" page for users with those roles.

I've based the design on the similar functionality in the
`app/views/account/organisations/edit.html.erb` &
`app/views/devise/invitations/new.html.erb` templates which have already
been converted to use the GOV.UK Design System.

I'm using the `error_summary` component in combination with identical
code we've used elsewhere so that the errors in the summary link to the
relevant form field. Similarly I'm using the `select` component with
`error_message` set appropriately, so that any relevant errors are
displayed alongside the `organisation_id` field and the field itself is
highlighted in red as per this Design System guidance [2]. I've enhanced
a test in `Users::OrganisationsControllerTest` to cover this new
behaviour. Note that in normal operation there are unlikely to be
validation errors on this page. However, it is possible if someone hacks
the form or e.g. you try to remove the organisation from an Organisation
Admin.

The button text has changed from "Update User" -> "Change organisation"
and I've introduced a "Cancel" link to match other similar pages.

[1]: https://github.com/alphagov/signon/commit/f71a128b6a6f9cd86bd8c08508a35d8780c128ef
[2]: https://design-system.service.gov.uk/patterns/validation/#how-to-tell-the-user-about-validation-errors