Simplify Users::OrganisationsController#user_params
c.f. this commit [1] where we did the same for `Users::NamesController`.

I found the `UserParameterSanitiser` and associated logic that we
brought from the `UsersController` quite confusing and somewhat overkill
for this controller. This commit removes the use of
`UserParameterSanitiser` but retains the use of `.permitted_user_params`
on the `Role::Base` subclasses. Note that only
`Roles::Superadmin.permitted_user_params` &
`Roles::Admin.permitted_user_params` return an array including
`:organisation_id` which is the only parameter we're interested in in
this controller.

I think ideally we'd make use of Pundit's strong parameters
functionality [2] and move this logic out of the `Roles::Base`
subclasses into the relevant policy class. However, that's a job for
another day!

[1]: https://github.com/alphagov/signon/commit/2c8010c059768774257c61cf14a04f515a9a0a14
[2]: https://github.com/varvet/pundit/blob/4d8cdf1c10058c12f2c175f30b408f5d5532a00b/README.md#strong-parameters