Consider target user in UserApplicationPermissionPolicy
TODO: Move the delete method logic to the SigninPermissionPolicy.

In preparation for moving this to the top-level of Pundit policies so
that we can use it when we introduce the ability to edit other users'
permissions.

Note the change of behaviour. A govuk admin no longer has to have access
to the app in order to be able to delete a UserApplicationPermission.