Use user instance variable vs f.object in form_fields partial
Previously we were using a mixture of these in
`app/views/users/_form_fields.html.erb` which I found confusing. The `f`
form builder is only really useful for creating form elements and we
won't be able to use those when we move the user edit page (and thus
this partial) to use the GOV.UK Design System components. So I think it
makes sense to use the instance variable consistently here.