Move policy check nearer functionality in user form_fields
When the `UserPolicy#mandate_2sv?` condition was first introduced [1],
although it was called `#flag_2sv?` at that point [2], it only wrapped
the `require_2sv` checkbox. This makes sense to me, because its that
checkbox that implements the mandating 2SV operation.

However, since then a bunch of extra 2SV-related functionality has been
included inside the condition, none of which is actually related to
mandating 2SV. I think this has happened accidentally or due to a
misunderstanding. So in this commit I have moved the condition as close
as possible to the `require_2sv` checkbox which makes more sense to me.

This won't currently make any difference, because
`UserPolicy#mandate_2sv?` is aliased to `UserPolicy#edit?` which is
required to be true for the whole page. However, I think this change
makes things less confusing.

This diff is best viewed with the `--ignore-all-space` option, because
a bunch of indentation has changed.

[1]: https://github.com/alphagov/signon/commit/0a8cc4df3737420b2480af8344efd40f1e97d731
[2]: https://github.com/alphagov/signon/commit/74b25a7191123a70f5d1d9ab4c7606fef48ffe57