Rename initiating_superadmin -> initiator in User#reset_2sv!
In this commit [1], `UserPolicy#reset_2sv?` was changed from only
allowing Superadmins to reset another user's 2SV to alias to
`UserPolicy#edit?` which is more permissive and allows Admins,
Organisation Admins & Super-Organisation Admins to reset _some_ user's
2SV. Thus the user passed in to `User#reset_2sv!` is no longer always a
Superadmin and so the `initiating_superadmin` argument name was
confusing.

Furthermore there wasn't anything in `User#reset_2sv!` to check that
`initiating_superadmin` was actually a Superadmin which was also
confusing.

[1]: https://github.com/alphagov/signon/commit/4d1c86d6a574f997c16e3ab46bceb49071851dde