Remove pointless tests for unsetting User#require_2sv
The assertions in these tests were passing even if all of the UI
interactions were removed. This was because they were all operating on
either the `user` or `user_in_child_organisation` instance variables
both of which already had `User#require_2sv` set to `false`, so
unsetting it is a no-op.

Furthermore, I couldn't fix the tests by setting `User#require_2sv` to
`true` in the setup, because then the checkbox for `User#require_2sv` in
`app/views/users/_form_fields.html.erb` did not show up and so the call
to `uncheck "Mandate 2-step verification for this user"` failed because
it couldn't find the checkbox.

In summary, these tests weren't testing what they were intended to test
and the behaviour they were intended to test wasn't available via the
user interface.

Note that since this commit [1], the only way to use the user interface
to unset `User#require_2sv` is to add an exemption for the user and this
is already tested elsewhere in `ManagingTwoStepVerificationTest`.

[1]: https://github.com/alphagov/signon/commit/a5bff442dee387087c9b72cc4a0d6d3286aa17d8