Update copy for UserMailer#reset_password_instructions
Previously we had a HTML link in the body of this email, but that is
not very accessible as it contains a token and screen readers read it
out in full. Instead we're relying on Notify's markdown support[1] to
turn this into a link and I've had to change the test helper which
parses out this URL in integration tests as a consequence.

[1] https://www.notifications.service.gov.uk/using-notify/links-and-URLs