Extract page for resending invitation for a user
Trello: https://trello.com/c/K3oa6wAj

This is a step on the journey to move the edit user page to use the
GOV.UK Design System. The new design calls for separate pages for
different operations on a user and this is the next one.

The new `app/views/users/invitation_resends/edit.html.erb` template is
based on other similar pages which we've recently extracted, e.g.
`app/views/users/two_step_verification_resets/edit.html.erb` where there
aren't any form fields; just a button.

Previously the "Resend signup email" button on the "edit user" page
submitted a POST request directly to `InvitationsController#resend`. Now
I've moved `InvitationsController#resend` to
`Users::InvitationResendsController#update` and a link on the "edit
user" page takes you to a new page which has a "Resend signup email"
button which submits a request to that new action.

I did consider adding an action to display the new page to the existing
`InvitationsController`. However, the latter is already significantly
changing the default implementation of `Devise::InvitationsController`
and I thought it would be better to move `InvitationsController#resend`
into the new controller. I think this also simplifies the
invitation-related routes a bit.

There weren't any tests in `UsersControllerTest` for this behaviour, so
I had to write `Users::InvitationResendsControllerTest` from scratch,
but I based it closely on other similar controller tests.  Luckily there
was one integration test (`test/integration/inviting_users_test.rb`)
covering the behaviour which I've been able to fixup to work with the
new UI.

Rather than creating a combinatorial explosion of tests in
`Users::InvitationResendsControllerTest` relating to whether a user with
all the different roles can resend an invitation to another user with
all the different roles, I've resorted to stubbing `UserPolicy.new` and
`UserPolicy#resend_invitation?`. Although this is a bit ugly, since
`UserPolicy` is thoroughly tested in `UserPolicyTest`, it seems like a
pragmatic option.

Since the `Users::InvitationResendsController#update` action calls a
bang method on the instance of `User` and doesn't check for errors,
there's no need to include an error summary component in the page.

The copy on the new "Resend signup email" page only really makes sense
in the context of a user who has been invited but hasn't yet accepted
the invitation and indeed the link to this page will only be displayed
on the "edit user" page in this scenario, so I've added a before action
to redirect back to the "edit user" if someone tries to view the page
for a user who has already accepted their invitation.