Update User#invited_by when resending signup email
This is a bit orthogonal to the current PR. However, it seemed
odd/inconsistent that previously we passed a value for `invited_by` to
`User.invite!` when the user was first invited, but we did not pass a
value for `invited_by` to `User#invite!` (i.e. the instance method) when
we resent an invitation. In fact it turns out that not passing
`invited_by` means that `User#invited_by` gets set to `nil`. Thus not
only did it not overwrite the initial value with an appropriate new
value, the original value was lost, which seems like the worst of all
worlds!