Record account invitation event on resending signup email
This is a bit orthogonal to the current PR. However, it seemed
odd/inconsistent that previously we recorded an account invitation event
when a user was first invited, but not if we resent the invitation.

It might be worth considering adding a new event type for this to
distinguish re-sending from the initial send, but I decided it wasn't
worthwhile right now.