Extract page for unlocking a user account
Trello: https://trello.com/c/P1w2E6MO

This is a step on the journey to move the edit user page to use the
GOV.UK Design System. The new design calls for separate pages for
different operations on a user and this is the next (and possibly last!)
one.

The new `app/views/users/unlockings/edit.html.erb` template is based on
other similar pages which we've recently extracted, e.g.
`app/views/users/invitation_resendings/edit.html.erb` where there aren't
any form fields; just a button.

Previously the "Unlock account" button on the "edit user" page submitted
a POST request directly to `UsersController#unlock`. Now I've moved
`UsersController#unlock` to `Users::UnlockingsController#update` and a
link on the "edit user" page takes you to a new page which has a "Unlock
account" button which submits a request to that new action.

There weren't any tests in `UsersControllerTest` for this behaviour, so
I had to write `Users::UnlockingsControllerTest` from scratch, but I
based it closely on other similar controller tests. Luckily there was
one integration test (`test/integration/user_locking_test.rb`) covering
the behaviour which I've been able to fixup to work with the new UI.

Rather than creating a combinatorial explosion of tests in
`Users::UnlockingsControllerTest` relating to whether a user with all
the different roles can resend an invitation to another user with all
the different roles, I've resorted to stubbing `UserPolicy.new` and
`UserPolicy#unlock?`. Although this is a bit ugly, since `UserPolicy` is
thoroughly tested in `UserPolicyTest`, it seems like a pragmatic option.

Since the `Users::UnlockingsController#update` action calls a bang
method on the instance of `User` and doesn't check for errors, there's
no need to include an error summary component in the page.

The copy on the new "unlock account" page only really makes sense in the
context of a user whose account has been locked and indeed the link to
this page will only be displayed on the "edit user" page in this
scenario, so I've added a before action to redirect back to the "edit
user" if someone tries to view the page for a user whose account is not
locked.

Note that I've used "unlockings" vs "unlocks" in the routes & controller
name to avoid clashing with the existing Devise-generated routes for a
user to unlock their own account, even though we don't include links to
this in the unlock instructions email, so they are currently unused. It
might make more sense to see if we can disable them and then rename the
routes & controller added in this commit, but I'm going to leave that
for now.