Use User.unlock_in in locked account explanations
This method is provided by Devise's Lockable module and reads its value
from the Devise configuration. The value is returned as an
`ActiveSupport::Duration` and so it can be converted into a
human-readable string by calling `ActiveSupport::Duration#inspect`.

Thus these changes make the copy resilient to changes in that
configuration and avoid a small amount of duplication.

Note that these changes won't currently change what is displayed,
because `Devise.unlock_in` is currently configured to be `1.hour`.