Add Users::SigninPermissionsController#create
This adds the `POST /users/<id>/applications/<id>/signin_permission`
route.

In order to know whether a user can grant another user access to an app,
we need to know about the current user, the target user and the
application. The UserApplicationPermission model already joins a User
and Application so we're defining our Pundit policy against this. Note
that we're somewhat abusing the UserApplicationPermission model by
instantiating it with a User and Application. In every other use of this
model we instantiate it with a User and a SupportedPermission, and the
Application is merely used as an optimisation. I've introduced the
`UserApplicationPermission.for` method in an attempt to differentiate
this use from the more standard construction of these models.

Note that Publishing Managers can only add the signin permission for a
user if both users have the signin permission and the signin permission
is delegatable. This duplicates the logic in
`UserPermissionManageableApplicationPolicy::Scope#resolve`[1] which is
currently used to determine the applications that appear in the
"Editable Permissions" table on the users edit page.

We'll update the user interface to make use of this new action in a
subsequent commit.

We've moved `stub_policy` into the new `PunditHelpers` module so that we
can use it in multiple tests.

[1]: https://github.com/alphagov/signon/blob/3282d8d0ce75acbb09feb5751e43f618846c1251/app/policies/user_permission_manageable_application_policy.rb#L25

Co-authored-by: Chris Lowis <chris.lowis@gofreerange.com>