Add Users::PermissionsController#show
This adds the `GET /users/<id>/applications/<id>/permissions` route
which lists each application permission along with whether the user has
been granted that permission.

Copied from `Account::PermissionsController#show` and updated to allow
the signed in user to view another user's permissions.

These permissions are currently visible on the user edit page
(/users/<id>/edit) which is only accessible to signed in users that can
edit the target user based on the logic in `UserPolicy#edit?`. We're
doing the same thing in this new action.

Note the additional `stub_policy_for_navigation_links` method in the
tests. This is to satisfy the call to `policy(User).index?` in
`NavigationItemsHelper#navigation_items`[1].

[1]: https://github.com/alphagov/signon/blob/2bd159ff90fcf293bf494425fc7653c263c2420a/app/helpers/navigation_items_helper.rb#L10

Co-authored-by: Chris Lowis <chris.lowis@gofreerange.com>