Split manage permissions page out of edit user page
Trello: https://trello.com/c/uZD2I9dj

This is (yet!) another step on the way to moving the edit user page to
use the GOV.UK Design System.

Work on moving the permission management functionality is ongoing [1],
but we are now in a position to move the rest of the edit user page to
use the Design System. Splitting the manage permissions form out means
we can e.g. change the layout for the edit user page.

I did consider moving the non-permission-related functionality into a
new `UsersController#show` action template and leaving the
permission-related functionality in the existing `UsersController#edit`
action template.  This made some sense, because we no longer need a form
for the non-permission-related functionality. However, it would have
meant a much more pervasive change, because the path to the edit user
page is used in many more places. Also even though there isn't a form,
it feels as if admins are most likely to visit the edit user page in
order to *edit* the user!

I've introduced a new policy predicate method,
`UserPolicy#manage_permissions?`, for the new action which is just
aliased to `UserPolicy#edit?`.

In `UsersControllerTest` I created a new context for the new
`manage_permissions` action and then either moved or copied tests from
the existing `edit` action as appropriate.

One minor fly in the ointment is that I've had to change the
`UsersController#update` action to render the new
`UsersController#manage_permissions` action template in the case of a
validation error. This is fine for when an admin is managing a user's
permissions, but it might be confusing when they are mandating 2SV for a
new user via the `UsersController#require_2sv` action template which
also submits to `UsersController#update`. However, it feels extremely
unlikely that there would ever be a validation error in the latter case
and even if there was, they user would at least see the error even if
they ended up on a surprising looking page! All this just makes me think
it's worth combining `UsersController#require_2sv` into
`Users::TwoStepVerificationMandationsController#edit` as I previously
suggested in #2547, but I'm going to leave that for now.

[1]: https://trello.com/c/baXtt5Dg