Hide suspend link unless current user has permission
c.f. 62609d2c2ae38620920701b979f3ffda0198328f

The `SuspensionsController` requires the current user to have the
`UserPolicy#suspension?` permission and so it makes sense that we only
display the link on the edit user page if the current user has that
permission.

Since `UserPolicy#suspension?` is currently aliased to
`UserPolicy#edit?` this doesn't make any practical difference, but since
we've defined `UserPolicy#suspension?` I think it makes sense to add
this condition in case the policy implementation changes in the future.