Hide suspend link for API user unless current user has permission
c.f. b8bdcf9bb9210e23c4cf8b76733ec8e9ea03ae2d

The `SuspensionsController` requires the current user to have the
`ApiUserPolicy#suspension?` and so it makes sense that we only display
the link on the edit API user page if the current user has that
permission.

Since `ApiUserPolicy#suspension?` is currently aliased to
`ApiUserPolicy#new?` this doesn't make any practical difference, but
since we've defined `ApiUserPolicy#suspension?` I think it makes sense
to add this condition in case the policy implementation changes in the
future.