Use User instance in all edit user permission checks
I'm not sure why these two checks were previously passing the `User`
class rather than the `user` instance to the `policy` helper method. I
suspect it might have been because neither of the `UserPolicy` predicate
methods actually depend on the user being edited; only on the current
user which is provided implicitly.

Anyway, I think it's clearer and more robust to consistently pass in the
`user` instance in case the policy implementation changes.