Move status to user details section on edit user page
Trello: https://trello.com/c/uZD2I9dj

This is a small step towards moving this page to the latest design which
will use the GOV.UK Design System. I'm trying to make small changes to
the page which will make the transition easier while retaining all the
functionality and keeping all the tests passing.

I've used an HTML description list element, because this most closely
resembles the Summary List component [1] that we're planning to use.

[1]: https://components.publishing.service.gov.uk/component-guide/summary_list