Move access log link to actions section on edit API user page
This is a small step towards moving this page to the latest design which
will use the GOV.UK Design System. I'm trying to make small changes to
the page which will make the transition easier while retaining all the
functionality and keeping all the tests passing.

Note that link is unchanged and it takes you to the existing access log
page which is implemented by `UsersController#event_logs` and works for
both users and API users. I've made a slight tweak to the
`title_caption` to distinguish the context from when you are viewing log
for a non-API user.

c.f. 2afff6e6aa1fbc1fee65ff3ab37ffdee661476bb