Move 2SV status to user details section on edit user page
Trello: https://trello.com/c/uZD2I9dj

This is a small step towards moving this page to the latest design which
will use the GOV.UK Design System. I'm trying to make small changes to
the page which will make the transition easier while retaining all the
functionality and keeping all the tests passing.

We're no longer displaying the 2SV exemption reason if an exemption
exists. Joseph has confirmed he's happy with this given the reason is
visible (on another page) if you edit the exemption.

The logic for the 2SV status is very slightly different compared with
that in `User#two_step_status` in that `User#has_2sv?` is now checked
*before* `User#exempt_from_2sv?`. I don't think this should make a
material difference and in any case I think it's better than we use the
shared `User#two_step_status` method rather than implementing our own
custom version. If there's a problem with the latter then that needs
fixing separately.