Move suspension link to actions section on edit API user page
This is a small step towards moving this page to the latest design which
will use the GOV.UK Design System. I'm trying to make small changes to
the page which will make the transition easier while retaining all the
functionality and keeping all the tests passing.

Note that link is unchanged and it takes you to the existing
suspend/unsuspend user page which is implemented by
`SuspensionsController` and works for both users and API users. I've
made a slight tweak to the `title_caption` to distinguish the context
from when you are acting on a non-API user.

c.f. e1c82a317bc17a6b912855aa67413dc03f16923f