Remove suspension reason from edit user page
Trello: https://trello.com/c/uZD2I9dj

This is a small step towards moving this page to the latest design which
will use the GOV.UK Design System. I'm trying to make small changes to
the page which will make the transition easier while retaining all the
functionality and keeping all the tests passing.

We're no longer displaying the suspension reason if a user is suspended
since the reason is visible (on another page) if you edit the
suspension.