Add Users::ApplicationsController#index
This adds the `GET /users/<id>/applications` route which displays the
applications that the user does and doesn't have access to.

Copied from Account::ApplicationsController and modified to allow the
signed in user to view the applications that another user has access to.

The signed in user can only view this page if they're allowed to `edit`
the other user according to the `UserPolicy`. This retains the behaviour
of the user edit page (/users/<id>/edit) which currently shows the list
of applications a user has access to.

Co-authored-by: Chris Lowis <chris.lowis@gofreerange.com>