Extract User#organisation_name method
This removes some duplication and I think it makes the code more
readable.

I also contemplated using it in `UserExportPresenter#row` instead of
`user.organisation.try(:name)`, but that would have changed the format
of the CSV, so I've left it alone for now.

Ideally I think we'd create a null Organisation object to further
simplify the code, but that can wait for another day.