Extract UsersHelper#link_to_resend_invitation
Note that this changes the behaviour very slightly, because if the
current user doesn't have permission to resend the signup email or if
the user being edited has accepted their invitation, then an *empty*
list item will be rendered. However, I have a plan on how to tackle that
in a subsequent commit when we start using the Summary List
component [1].

I decided to remove a bunch of combinatorial tests in
`UsersControllerTest`, because the new tests in `UsersHelperTest` for
`UsersHelper#link_to_resend_invitation` provide better coverage of the
same logic.

[1]: https://components.publishing.service.gov.uk/component-guide/summary_list