Extract UsersHelper#link_to_unlock
Note that this changes the behaviour very slightly, because if the
current user doesn't have permission to unlock the user or if the user
being edited is not locked, then an *empty* list item will be rendered.
However, I have a plan on how to tackle that in a subsequent commit when
we start using the Summary List component [1].

I decided to remove a bunch of combinatorial tests in
`UsersControllerTest`, because the new tests in `UsersHelperTest` for
`UsersHelper#link_to_unlock` provide better coverage of the same logic.

[1]: https://components.publishing.service.gov.uk/component-guide/summary_list