Move edit user page to use GOV.UK Design System 🎉
This introduces the Summary List component [1] that we've been working
towards for displaying the attributes of the user along with links to
change them. I've extracted methods into `UsersHelper` to generate the
appropriate options for each of the items. I decided to replace a bunch
of combinatorial tests in `UsersControllerTest` with unit tests for
these new helper methods.

The links in the "Actions" section are displayed using a List
component [2] and its items are obtained from a bunch of
`UsersHelper#link_to_*` methods. As I mentioned when I added the latter,
these return `nil` if the link is not to be displayed and we handle
that here by compacting the array of items before passing it to the
component.

[1]: https://components.publishing.service.gov.uk/component-guide/summary_list
[2]: https://components.publishing.service.gov.uk/component-guide/list