Extract User#signin_permission_for
To DRY up controllers and templates.

We were previously relying on the call to `find_by!` in the controllers
to raise an exception if the user didn't have the signin permission for
the application. This is now handled by `set_application` in both
controllers so it's OK that `User#signin_permission` returns nil if it
doesn't exist.

Co-authored-by: Chris Lowis <chris.lowis@gofreerange.com>